Modified searchState array and animate()
Changed the usage of the searchState array to store only the index
values of the slice of the data array that are still being searched.

Modified animate() to handle the new format of the searchState array.
Discovered some coding mistakes in there that somehow didn't break the
script before.

Made some more descriptive console.logs.